Changed the PCL Profile to Profile78 (.NET 4.5, WinRT, WinPhone)
Silverlight can't use unsafe code, so unless we rewrite all of the
encodings to not use pointers, we can't target Silverlight.

Since we can't target Silverlight without a lot of effort, might
as well target Profile78 which will allow us to remove the ToCharArray()
hack which should greatly improve performance.